Move NumPy specific types to NumPy module
